Implement top-10 codebase/UX review recommendations
- Onboarding: consent card now states the auto-record default (was
  "starts in manual mode"); the onboarding-shown flag persists on
  explicit finish/defer instead of first boot, so a mid-wizard quit
  re-opens setup. Fully-permissioned installs are still recognised.
- Sidebar privacy footer is state-aware: with a remote Think backend it
  shows an amber dot and "Think uses an approved remote server" instead
  of the unconditional "No data leaves your Mac" (AppSettings gains
  usesRemoteMainLLM).
- Naming unified: Ask (not "Conversations"), Autocomplete (not
  "Cotyping") in all user-facing strings, Day digest (not "Day brief")
  across Timeline; empty-state copy loses the "persistent header"
  dev-speak.
- Failure recovery: Retry buttons on meeting detail and workspace
  failure states; DiskSpacePrecheck gains a recording floor and
  RecordingController refuses to start on a nearly-full volume.
- Deleted dead CotypingView.swift (unreachable ~22-control settings
  clone; tuning lives in Settings, rehearsal in Type).
- Timeline: Time allocation now excludes system-only apps
  (loginwindow…); session titles strip browser chrome suffixes
  ("… - Audio playing - Google Chrome - Profile").
- File split: AppState out of LokalBotApp.swift; MeetingDetailView,
  MarkdownText and GettingStartedCard out of MainWindowView.swift.
- Tokens: semantic Brand.error replaces hardcoded .orange error text,
  Ask CTA runs through Brand.teal, StatusDot honors Reduce Motion.
- Policy tightening: 0.0.0.0 no longer counts as loopback for inference
  endpoints; model downloads must be https. AgentAccessManager surfaces
  enable failures to the error toast instead of swallowing them.
- New tests: RecordingController, LlamaServer, usesRemoteMainLLM,
  recording disk floor, plaintext-download refusal, 0.0.0.0 policy,
  browser-chrome title stripping. Full unit suite: 1494 passed.
