+++ LokalBot/AppState.swift
@@ -1,0 +1,1440 @@
+import SwiftUI
+import Combine
+import AVFoundation
+import CoreGraphics
+
+/// Central app state (the "coordinator" from the design doc §6): dependency
+/// wiring, the meeting library, navigation, and the detection→recording glue.
+/// The recording lifecycle itself lives in `RecordingController`; headless
+/// subcommands in `HeadlessCommandRunner`; AppKit launch plumbing in
+/// `AppLifecycle.swift`.
+@MainActor
+final class AppState: ObservableObject {
+
+    enum NavSection: Hashable {
+        case today, timeline, meetings, type, ask, agent, settings
+
+        /// Section names accepted from the UI-test capture environment and
+        /// deep links. Legacy names keep working: "capture" (the pre-split
+        /// merged section) lands on Timeline, "dictation" and "cotyping" on
+        /// Type, "search"/"chat" on Ask (spec §2.1), and "models" on
+        /// Settings, which absorbed it as a tab (spec §2.5).
+        init?(captureName: String) {
+            switch captureName.lowercased() {
+            case "today": self = .today
+            case "timeline", "capture": self = .timeline
+            case "meetings": self = .meetings
+            case "type", "dictation", "cotyping", "autocomplete": self = .type
+            case "ask", "search", "chat": self = .ask
+            case "agent": self = .agent
+            case "settings", "models": self = .settings
+            default: return nil
+            }
+        }
+    }
+
+    /// Which tool the Type section shows. Session-sticky: preserved while
+    /// navigating away so returning lands on the last-used tab.
+    enum TypeTab: String, CaseIterable {
+        case dictation, cotyping
+
+        var displayName: String {
+            self == .cotyping ? "Autocomplete" : "Dictation"
+        }
+
+        /// Legacy capture names select their tab; anything else leaves the
+        /// current tab untouched.
+        init?(captureName: String) {
+            switch captureName.lowercased() {
+            case "dictation": self = .dictation
+            case "cotyping", "autocomplete": self = .cotyping
+            default: return nil
+            }
+        }
+    }
+
+    enum MeetingPresentation: String {
+        case preview
+        case detail
+    }
+
+    struct AgentLaunchContext: Equatable {
+        var title: String
+        var prompt: String
+        var meetingID: Meeting.ID?
+        var actionID: String?
+    }
+
+    /// Which tab the Settings surface shows (spec §2.5 — Settings absorbs
+    /// Models as a tab strip). Session-sticky like TypeTab.
+    enum SettingsTab: String, CaseIterable {
+        case general, recording, models, privacy, advanced
+
+        var displayName: String { rawValue.capitalized }
+
+        /// Legacy capture names select their tab; the pre-merge "models"
+        /// section name lands on the Models tab.
+        init?(captureName: String) {
+            switch captureName.lowercased() {
+            case "general": self = .general
+            case "recording": self = .recording
+            case "models": self = .models
+            case "privacy": self = .privacy
+            case "advanced": self = .advanced
+            default: return nil
+            }
+        }
+    }
+
+    /// True when launched by an XCUITest harness — gates the side-effectful
+    /// startup paths (Core Audio polling, accessibility-trusted detector,
+    /// Sparkle, periodic screenshots) so the UI renders against synthetic
+    /// data without touching real audio, TCC, or the network.
+    nonisolated static var isUITesting: Bool { UITestRuntime.isEnabled }
+    /// Hosted XCTest executes the real app entry point. It needs neither UI
+    /// fixtures nor any interactive service, so return before touching the
+    /// meeting library, indexes, permissions, audio, Sparkle, or the network.
+    nonisolated static var isUnitTesting: Bool { UITestRuntime.isUnitTesting }
+
+    /// UserDefaults flag: the permission onboarding has been shown once. Also
+    /// read by `AppDelegate` to keep the first run windowed.
+    nonisolated static let onboardingShownKey = "lokalbotv3.onboarding.shown"
+
+    @Published private(set) var meetings: [Meeting] = []
+    @Published var lastError: String?
+    /// A recording or dictation start was refused because microphone access
+    /// is denied at the system level. Cleared when the user opens System
+    /// Settings from the recovery toast or dismisses it.
+    @Published var micRecoveryNeeded = false
+
+    /// The always-alive settings owner; services capture this, never AppState.
+    let settingsStore = SettingsStore()
+
+    /// UI-facing settings surface (views bind `$app.settings.x`). Writes flow
+    /// through to `settingsStore` (which persists) and fan out to the
+    /// subsystems that apply settings live.
+    @Published var settings: AppSettings {
+        didSet {
+            guard settings != oldValue else { return }
+            settingsStore.current = settings
+            if settings.stopDebounceSeconds != oldValue.stopDebounceSeconds {
+                detector.stopDebounce = settings.stopDebounceSeconds
+            }
+            if settings.calendarDetectionEnabled != oldValue.calendarDetectionEnabled {
+                detector.calendarEnabled = settings.calendarDetectionEnabled
+            }
+            if settings.requireCalendarForBrowser != oldValue.requireCalendarForBrowser {
+                detector.requireCalendarForBrowser = settings.requireCalendarForBrowser
+            }
+            if interactive {
+                if Self.dictationLifecycleChanged(from: oldValue, to: settings) {
+                    dictation.applySettings()
+                }
+                if settings.cotypingEnabled != oldValue.cotypingEnabled {
+                    cotyping.applySettings()
+                    if !settings.cotypingEnabled {
+                        scheduleCotypingRuntimeUnload()
+                    }
+                }
+                if Self.cotypingRuntimeChanged(from: oldValue, to: settings) {
+                    scheduleCotypingPrewarm()
+                }
+                if settings.quickRecallEnabled != oldValue.quickRecallEnabled {
+                    applyQuickRecallSetting()
+                }
+                if ScreenshotRetentionSchedule.requiresImmediatePrune(
+                    previousDays: oldValue.retentionDays,
+                    currentDays: settings.retentionDays
+                ) {
+                    screenshots.pruneOldScreenshots()
+                }
+                if Self.dailyMemoryExportChanged(from: oldValue, to: settings) {
+                    applyDailyMemoryExportSetting()
+                }
+                if Self.dayDigestChanged(from: oldValue, to: settings) {
+                    applyDayDigestSetting()
+                }
+                if Self.screenContextChanged(from: oldValue, to: settings) {
+                    applyTrackingSetting()
+                }
+                if Self.memoryRoutinesChanged(from: oldValue, to: settings) {
+                    applyMemoryRoutineSetting()
+                }
+                if Self.dreamingChanged(from: oldValue, to: settings) {
+                    applyDreamingSetting()
+                }
+            }
+        }
+    }
+
+    private static func dictationLifecycleChanged(from old: AppSettings, to new: AppSettings) -> Bool {
+        old.dictationEnabled != new.dictationEnabled
+            || old.dictationShowOverlay != new.dictationShowOverlay
+            || old.dictationLivePreview != new.dictationLivePreview
+    }
+
+    private static func cotypingRuntimeChanged(from old: AppSettings, to new: AppSettings) -> Bool {
+        guard new.cotypingEnabled else { return false }
+        return !old.cotypingEnabled
+            || old.cotypingBuiltInModelID != new.cotypingBuiltInModelID
+            || old.cotypingInProcessRuntime != new.cotypingInProcessRuntime
+            || old.customBuiltInModels != new.customBuiltInModels
+    }
+
+    private static func dailyMemoryExportChanged(from old: AppSettings,
+                                                 to new: AppSettings) -> Bool {
+        old.dailyMemoryExportEnabled != new.dailyMemoryExportEnabled
+            || old.dailyMemoryExportFolder != new.dailyMemoryExportFolder
+            || old.dailyMemoryExportFormat != new.dailyMemoryExportFormat
+            || old.dailyMemoryExportHour != new.dailyMemoryExportHour
+    }
+
+    /// The custom prompt is deliberately absent: it is read fresh from
+    /// `settings` at generation time, so editing it never resets the scheduler.
+    private static func dayDigestChanged(from old: AppSettings,
+                                         to new: AppSettings) -> Bool {
+        old.dayDigestAutoEnabled != new.dayDigestAutoEnabled
+            || old.dayDigestHour != new.dayDigestHour
+    }
+
+    private static func screenContextChanged(from old: AppSettings,
+                                             to new: AppSettings) -> Bool {
+        old.trackingEnabled != new.trackingEnabled
+            || old.effectiveScreenContextCaptureMode != new.effectiveScreenContextCaptureMode
+            || old.screenshotIntervalMinutes != new.screenshotIntervalMinutes
+            || old.meetingVisualContextEnabled != new.meetingVisualContextEnabled
+    }
+
+    private static func memoryRoutinesChanged(from old: AppSettings,
+                                              to new: AppSettings) -> Bool {
+        old.memoryRoutinesEnabled != new.memoryRoutinesEnabled
+            || old.memoryRoutineFolder != new.memoryRoutineFolder
+            || old.enabledMemoryRoutines != new.enabledMemoryRoutines
+            || old.memoryRoutineHour != new.memoryRoutineHour
+            || old.memoryRoutineWeekday != new.memoryRoutineWeekday
+    }
+
+    private static func dreamingChanged(from old: AppSettings,
+                                        to new: AppSettings) -> Bool {
+        old.dreamingEnabled != new.dreamingEnabled
+            || old.dreamingHour != new.dreamingHour
+            || old.dreamingFirstEligibleDayKey != new.dreamingFirstEligibleDayKey
+    }
+
+    // Navigation (main window): sidebar section, selected meeting, and a
+    // pending "jump to timestamp" handed from search to the detail player.
+    @Published var navSection: NavSection = .today
+    private static let typeTabDefaultsKey = "lokalbotv3.type.selectedTab"
+    private static var navigationDefaults: UserDefaults {
+        if let suite = UITestRuntime.defaultsSuiteName,
+           let defaults = UserDefaults(suiteName: suite) {
+            return defaults
+        }
+        return .standard
+    }
+    @Published var typeTab: TypeTab = .dictation {
+        didSet { Self.navigationDefaults.set(typeTab.rawValue, forKey: Self.typeTabDefaultsKey) }
+    }
+    @Published var settingsTab: SettingsTab = .general
+    @Published var selectedMeetingIDs: Set<Meeting.ID> = []
+    @Published var meetingPresentation: MeetingPresentation = .preview
+    @Published var pendingSeek: TimeInterval?
+    /// A screen-memory deep link waiting for Timeline's shared capture model.
+    /// Search, assistant citations, and Quick Recall set this before switching
+    /// sections; Timeline consumes it once its content column is mounted.
+    @Published var pendingScreenSnapshotID: Int64?
+
+    /// A query handed to the Ask section by another surface (⌘K palette).
+    /// AskView consumes and clears it on appear/change.
+    @Published var askPrefill: String?
+
+    /// True when an explicitly ask-labelled handoff should send immediately,
+    /// rather than merely placing text in Ask's search field.
+    @Published var askSubmitRequested = false
+
+    /// A day handed to the Ask section (the old Timeline "Ask" tab, spec
+    /// §2.2): rendered as a removable chip, and prepended to escalated
+    /// queries so the assistant scopes its answer to that day.
+    @Published var askDayScope: Date?
+    /// Screens explicitly attached from Timeline to the next Ask turn.
+    @Published var askScreenContextIDs: [Int64] = []
+    /// Contextual handoffs only prefill Agent. The subprocess starts after the
+    /// user reviews the prompt and explicitly chooses Send.
+    @Published var agentLaunchContext: AgentLaunchContext?
+
+    /// Navigate to the Type section with a specific tab preselected.
+    func openType(_ tab: TypeTab) {
+        typeTab = tab
+        navSection = .type
+    }
+
+    /// Navigate to Settings with a specific tab preselected.
+    func openSettings(tab: SettingsTab) {
+        settingsTab = tab
+        navSection = .settings
+    }
+
+    /// Navigate to the Ask section, optionally pre-filling the query and/or
+    /// scoping it to a day (Timeline's "Ask about this day").
+    func openAsk(query: String = "", dayScope: Date? = nil,
+                 screenSnapshotIDs: [Int64] = [], submit: Bool = false) {
+        askPrefill = query.isEmpty ? nil : query
+        askDayScope = dayScope
+        askScreenContextIDs = screenSnapshotIDs
+        askSubmitRequested = submit
+        navSection = .ask
+    }
+
+    func openAgent(_ context: AgentLaunchContext? = nil) {
+        agentLaunchContext = context
+        if let prompt = context?.prompt {
+            agentSessions.ensureSelectedController().draft = prompt
+        }
+        navSection = .agent
+    }
+
+    /// Open one meeting in the Meetings section — the deep-link target
+    /// for search hits, menu-bar recents, and palette recents.
+    func openMeeting(_ id: Meeting.ID) {
+        selectedMeetingIDs = [id]
+        meetingPresentation = .detail
+        navSection = .meetings
+    }
+
+    func previewMeeting(_ id: Meeting.ID) {
+        selectedMeetingIDs = [id]
+        meetingPresentation = .preview
+        navSection = .meetings
+    }
+
+    func selectDefaultMeetingIfNeeded() {
+        guard selectedMeetingIDs.isEmpty else { return }
+        if let live = currentMeeting {
+            previewMeeting(live.id)
+        } else if let completed = meetings.first(where: { meeting in
+            meeting.endedAt != nil && pipeline.stages[meeting.id] == nil
+        }) ?? meetings.first(where: { $0.endedAt != nil }) {
+            previewMeeting(completed.id)
+        }
+    }
+
+    /// The meeting shown in the detail pane (single selection only). The
+    /// in-progress recording is a first-class citizen here: selecting its
+    /// row resolves to `currentMeeting`, which `CaptureDetailView` routes
+    /// to the live view.
+    var selectedMeeting: Meeting? {
+        guard selectedMeetingIDs.count == 1, let id = selectedMeetingIDs.first else { return nil }
+        if let live = currentMeeting, live.id == id { return live }
+        return meetings.first { $0.id == id }
+    }
+
+    let storage = StorageManager()
+    private(set) lazy var outcomeIndex = OutcomeIndex(storage: storage)
+    private(set) lazy var cotypingLearning = CotypingLearningStore(storageRoot: storage.rootURL)
+    let detector = MeetingDetector()
+    let audioMonitor = AudioSourceMonitor()
+    /// External-agent access: Privacy marker plus the local-model wake watcher.
+    private(set) lazy var agentAccess = AgentAccessManager(
+        storage: storage,
+        settings: { [weak self] in self?.settings ?? AppSettings.load() },
+        onError: { [weak self] message in self?.lastError = message })
+    /// Screen-memory access is intentionally separate from meeting-library
+    /// access: enabling one capability never grants the other.
+    private(set) lazy var screenMemoryAccess = ScreenMemoryAccessManager(
+        gate: ScreenMemoryAccessGate(root: storage.rootURL))
+    private(set) lazy var quickRecallHotKey: QuickRecallHotKeyController = {
+        let controller = QuickRecallHotKeyController()
+        controller.onInvoke = { WindowAccess.shared.open("quick-recall") }
+        return controller
+    }()
+    private let dailyMemoryExportScheduler = DailyMemoryExportScheduler()
+    private let dayDigestScheduler = DayDigestScheduler()
+    private(set) lazy var memoryRoutines = MemoryRoutineScheduler(
+        storageRoot: storage.rootURL,
+        databaseURL: storage.rootURL.appendingPathComponent("lokalbotv3.sqlite"),
+        canRun: { [weak self] in
+            guard let self else { return false }
+            let cotypingGenerating: Bool
+            if case .generating = self.cotyping.state {
+                cotypingGenerating = true
+            } else {
+                cotypingGenerating = false
+            }
+            return !self.recording.isRecording
+                && !self.recording.isStarting
+                && !self.dictation.isStarting
+                && !self.dictation.state.isWorking
+                && !cotypingGenerating
+                && !self.pipeline.hasActiveWork
+        })
+    /// Overnight dreaming: the scheduler is observed by Settings (progress /
+    /// last-run), the store is read by Today, and `latestDreamReport` lets an
+    /// open Today window pick up a dream that finishes while it's visible.
+    @Published var latestDreamReport: DreamReport?
+    /// Settings' user-managed pin controls. The store remains the source of
+    /// truth; mutations reload it before saving to avoid stale-view writes.
+    @Published private(set) var dreamMemory: DreamMemory?
+    private var dreamObserver: AnyCancellable?
+    private(set) lazy var dreamStore = DreamStore(root: storage.rootURL)
+    private(set) lazy var dreaming: DreamScheduler = {
+        let scheduler = DreamScheduler()
+        dreamObserver = scheduler.objectWillChange.sink { [weak self] in
+            self?.objectWillChange.send()
+        }
+        return scheduler
+    }()
+    /// Read-only calendar access (EventKit): confirms meetings and titles
+    /// recordings. Concrete type so the settings UI observes its permission
+    /// state; handed to the detector as the `CalendarEventProviding` seam.
+    let calendar = EventKitCalendarEventProvider()
+    private var cachedSearchIndex: SearchIndex?
+    var searchIndex: SearchIndex {
+        if let cachedSearchIndex { return cachedSearchIndex }
+        let created = SearchIndex(
+            databaseURL: storage.rootURL.appendingPathComponent("lokalbotv3.sqlite"))
+        for meetingID in deletedMeetingIDs { created.noteDeletion(meetingID) }
+        cachedSearchIndex = created
+        return created
+    }
+    private(set) lazy var activityStore = ActivityStore(
+        databaseURL: storage.rootURL.appendingPathComponent("lokalbotv3.sqlite"))
+    private(set) lazy var sampler = ActivitySampler(store: activityStore)
+    private var cachedEmbeddingIndex: EmbeddingIndex?
+    var embeddingIndex: EmbeddingIndex {
+        if let cachedEmbeddingIndex { return cachedEmbeddingIndex }
+        let created = EmbeddingIndex(
+            databaseURL: storage.rootURL.appendingPathComponent("lokalbotv3.sqlite"),
+            storage: storage)
+        for meetingID in deletedMeetingIDs { created.noteDeletion(meetingID) }
+        cachedEmbeddingIndex = created
+        return created
+    }
+    private(set) lazy var screenshots = ScreenshotService(
+        store: activityStore, storage: storage, sampler: sampler,
+        isMeetingRecordingActive: { [weak self] in
+            guard let self else { return false }
+            return self.recording.isRecording || self.recording.isStarting
+        }, activeMeetingID: { [weak self] in
+            self?.recording.currentMeeting?.id
+        }, isHighPriorityInteractionActive: { [weak self] in
+            guard let self else { return false }
+            if self.dictation.state.isWorking { return true }
+            if case .generating = self.cotyping.state { return true }
+            return false
+        }) { [store = settingsStore] in
+        store.current
+    }
+    private(set) lazy var pipelineJobStore = PipelineJobStore(
+        databaseURL: storage.rootURL.appendingPathComponent("lokalbotv3.sqlite"))
+    private(set) lazy var pipeline = ProcessingPipeline(
+        storage: storage, jobStore: pipelineJobStore) { [store = settingsStore] in
+        store.current
+    }
+    /// Meeting-recording lifecycle: recorders, watchdog, timer tick, prewarm.
+    private(set) lazy var recording = RecordingController(
+        storage: storage,
+        settingsStore: settingsStore,
+        audioMonitor: audioMonitor,
+        pipeline: pipeline,
+        isInteractive: { [weak self] in self?.interactive ?? false },
+        onError: { [weak self] message in self?.lastError = message },
+        onMicPermissionDenied: { [weak self] in self?.micRecoveryNeeded = true },
+        onMeetingFinished: { [weak self] meeting in self?.meetings.insert(meeting, at: 0) })
+    /// Press-and-speak composition. Every dictation is treated as a writing
+    /// request: local ASR captures the instruction, the configured composition
+    /// LLM writes against ephemeral focused-window context, and focus-safe
+    /// delivery inserts it.
+    private(set) lazy var dictation = DictationCoordinator(
+        storageRoot: storage.rootURL,
+        settingsProvider: { [store = settingsStore] in store.current },
+        makeTextEngine: { [weak self] in
+            guard let self else {
+                throw TextEngineError.unavailable("LokalBot is shutting down.")
+            }
+            let config = self.settingsStore.current.dictationCompositionTextEngineSettings
+            return try await self.pipeline.makeTextEngine(
+                config,
+                priority: .interactive,
+                purpose: "dictation compose")
+        },
+        canStart: { [weak self] in
+            guard let self else { return false }
+            return !self.recording.isRecording
+                && !self.recording.isStarting
+                && self.pendingRecordingStart == nil
+        },
+        onBusy: { [weak self] in
+            self?.lastError = "Stop the current meeting recording before starting dictation."
+        },
+        onError: { [weak self] message in
+            self?.lastError = message
+        },
+        onMicPermissionDenied: { [weak self] in
+            self?.micRecoveryNeeded = true
+        })
+    /// Cotyping (inline AI autocomplete). Always runs its own model on the
+    /// dedicated `LlamaServer.cotyping` instance so it never thrashes the
+    /// shared summarizer server. Resolved per-completion, so changes apply live.
+    private(set) lazy var cotypingEngine = CotypingEngineSelector(
+        http: CotypingEngine(
+            makeEngine: { [weak self] in
+                guard let self else { throw TextEngineError.unavailable("LokalBot is shutting down.") }
+                return try await self.pipeline.makeTextEngine(
+                    self.settings.cotypingTextEngineSettings,
+                    server: .cotyping,
+                    priority: .interactive,
+                    purpose: "cotyping")
+            },
+            stopRuntime: { await LlamaServer.cotyping.stop() }),
+        makeLocal: { modelPath in
+            LocalLlamaCotypingEngine(runtime: LlamaCotypingRuntime(), modelPath: modelPath)
+        },
+        settings: { [store = settingsStore] in store.current },
+        storage: storage)
+    private(set) lazy var cotyping = CotypingCoordinator(
+        engine: cotypingEngine,
+        settingsProvider: { [store = settingsStore] in store.current },
+        learningStore: cotypingLearning,
+        isMeetingRecordingActive: { [weak self] in
+            guard let self else { return false }
+            return self.recording.isRecording || self.recording.isStarting
+        })
+    /// Rolling live transcript of the recording in progress (preview only —
+    /// the pipeline's post-meeting transcript stays authoritative).
+    private(set) lazy var liveTranscriber = LiveMeetingTranscriber(
+        storageRoot: storage.rootURL,
+        settings: { [store = settingsStore] in store.current })
+
+    @MainActor
+    func prepareRecommendedCotypingModel() {
+        CotypingModelPreparer.prepareRecommended(
+            settings: &settings,
+            storage: storage,
+            downloads: ModelDownloadManager.shared)
+    }
+
+    /// Chat assistant (the "Chat" section). Reuses the summariser's `TextEngine`
+    /// and a tool-calling agent over the live meeting list + search indexes.
+    private(set) lazy var chat = ChatViewModel(
+        makeEngine: { [weak self] in
+            guard let self else { throw TextEngineError.unavailable("LokalBot is shutting down.") }
+            return try await self.pipeline.makeTextEngine(
+                self.settings,
+                priority: .interactive,
+                purpose: "chat")
+        },
+        tools: MeetingChatTools(
+            meetings: { [weak self] in self?.meetings ?? [] },
+            storage: storage,
+            searchIndex: searchIndex,
+            embeddingIndex: embeddingIndex,
+            activityStore: activityStore,
+            settings: { [store = settingsStore] in store.current }),
+        store: ChatStore(rootURL: storage.rootURL),
+        workMemory: { [weak self] in
+            guard let self else { return "" }
+            return ChatPrompt.workMemoryContext(
+                memory: (try? self.dreamStore.loadMemory()) ?? nil,
+                dreamingEnabled: self.settings.dreamingEnabled)
+        })
+
+    // Agent Mode (pi). Installer and tab manager are cheap; each tab lazily
+    // spawns its own controller/process when AgentView mounts it.
+    let agentInstaller = AgentRuntimeInstaller()
+    private(set) lazy var agentSessions = AgentSessionTabs(
+        settings: { [store = settingsStore] in store.current },
+        storage: storage)
+    var agentController: AgentSessionController {
+        agentSessions.ensureSelectedController()
+    }
+
+    private var pipelineObserver: AnyCancellable?
+    private var outcomeObserver: AnyCancellable?
+    private var dictationObserver: AnyCancellable?
+    private var recordingObserver: AnyCancellable?
+    private var recordingStatusObserver: AnyCancellable?
+    private var audioMonitorObserver: AnyCancellable?
+    private var audioMonitorChangeForwarder: AnyCancellable?
+    private var calendarObserver: AnyCancellable?
+    /// True only on the real interactive launch path (not headless / UI test) —
+    /// gates recording notifications and first-run onboarding.
+    private var interactive = false
+    private var terminationCleanupTask: Task<Void, Never>?
+    /// Serializes cotyping model lifecycle transitions. A disable must finish
+    /// unloading both runtimes before a rapid re-enable can prewarm a fresh
+    /// model, otherwise the old and new routes can overlap in memory.
+    private var cotypingRuntimeTask: Task<Void, Never>?
+    private var cotypingRuntimeTaskID: UUID?
+    private var libraryLoadTask: Task<Void, Never>?
+    private var embeddingIndexTasks: [Meeting.ID: (token: UUID, task: Task<Void, Never>)] = [:]
+    private var indexCleanupTasks: [Meeting.ID: (token: UUID, task: Task<Void, Never>)] = [:]
+    private var deletedMeetingIDs: Set<Meeting.ID> = []
+    @Published private(set) var libraryReady = false
+    private var pendingRecordingStart: (context: MeetingDetectionContext?, source: String)?
+    private lazy var searchIndexWorkQueue = SearchIndexWorkQueue(
+        databaseURL: storage.rootURL.appendingPathComponent("lokalbotv3.sqlite"),
+        rootURL: storage.rootURL)
+
+    // Recording facades — views observe AppState only.
+    var isRecording: Bool { recording.isRecording }
+    var currentMeeting: Meeting? { recording.currentMeeting }
+    var elapsed: TimeInterval { recording.elapsed }
+    var menuBarTimer: String { recording.menuBarTimer }
+
+    func startRecording(context: MeetingDetectionContext? = nil, source: String = "ui") {
+        RecordingNotifier.shared.invalidateMeetingDetections()
+        guard !dictation.isStarting, !dictation.state.isWorking else {
+            lastError = "Finish or cancel dictation before starting a meeting recording."
+            return
+        }
+        guard libraryReady else {
+            pendingRecordingStart = (context, source)
+            lastError = "Preparing your meeting library; recording will start when it is ready."
+            return
+        }
+        recording.start(context: context, source: source)
+    }
+
+    func stopRecording(process: Bool = true) {
+        if !libraryReady {
+            pendingRecordingStart = nil
+            return
+        }
+        recording.stop(process: process)
+    }
+
+    /// Recording started, stopped, or split to a new meeting: quiet cotyping
+    /// and (re)point the live transcriber at the active meeting folder.
+    private func meetingRecordingStateDidChange(active: Bool) {
+        guard interactive else { return }
+        if settings.cotypingEnabled {
+            cotyping.meetingRecordingStateChanged(active: active)
+        }
+        if active, let meeting = recording.currentMeeting {
+            liveTranscriber.prepare(folder: meeting.folderURL(in: storage))
+        } else {
+            liveTranscriber.stop()
+        }
+    }
+
+    /// Menu-bar path to the live meeting: land on the library with the
+    /// in-progress recording selected, so the detail pane shows the live
+    /// transcript and notes.
+    func showLiveMeeting() {
+        guard let live = currentMeeting else { return }
+        navSection = .meetings
+        selectedMeetingIDs = [live.id]
+        WindowAccess.shared.open("main")
+    }
+
+    init() {
+        AppLog.bootstrap()
+        settings = settingsStore.current
+        if let raw = Self.navigationDefaults.string(forKey: Self.typeTabDefaultsKey),
+           let stored = TypeTab(rawValue: raw) {
+            typeTab = stored
+        } else {
+            // Existing installs retain Dictation; genuinely new installs lead
+            // with the approved Autocomplete experience.
+            typeTab = Self.navigationDefaults.bool(forKey: Self.onboardingShownKey)
+                ? .dictation : .cotyping
+        }
+        if Self.isUnitTesting { return }
+        let headlessCommand = HeadlessCommand.requested
+        let needsSynchronousLibrary = headlessCommand != nil || Self.isUITesting
+        if needsSynchronousLibrary {
+            meetings = storage.loadMeetings()
+            libraryReady = true
+            outcomeIndex.refresh(meetings: meetings)
+        }
+        LiveMeetingTranscriber.sweepOrphanedSnapshots(storageRoot: storage.rootURL)
+        // Views observe AppState only; forward pipeline / recording /
+        // audio-monitor / calendar change notifications so MainWindowView
+        // refreshes when those sub-ObservableObjects publish.
+        pipelineObserver = pipeline.objectWillChange.sink { [weak self] in
+            self?.objectWillChange.send()
+        }
+        outcomeObserver = outcomeIndex.objectWillChange.sink { [weak self] in
+            self?.objectWillChange.send()
+        }
+        dictationObserver = Publishers.MergeMany([
+            dictation.$state.dropFirst().map { _ in () }.eraseToAnyPublisher(),
+            dictation.$isStarting.dropFirst().map { _ in () }.eraseToAnyPublisher(),
+            dictation.$isShortcutMonitoringActive.dropFirst().map { _ in () }.eraseToAnyPublisher(),
+            dictation.$lastTranscript.dropFirst().map { _ in () }.eraseToAnyPublisher(),
+            dictation.$lastComposedText.dropFirst().map { _ in () }.eraseToAnyPublisher(),
+            dictation.$lastEngine.dropFirst().map { _ in () }.eraseToAnyPublisher(),
+        ]).sink { [weak self] in
+            self?.objectWillChange.send()
+        }
+        // Forward lifecycle changes, but not RecordingController's one-second
+        // `now` tick. Timer labels observe the controller directly; rebroadcasting
+        // that tick through AppState invalidated the entire window hierarchy.
+        recordingObserver = Publishers.CombineLatest(
+            recording.$status,
+            recording.$currentMeeting
+        ).dropFirst().sink { [weak self] _ in
+            self?.objectWillChange.send()
+        }
+        // React to recording start/stop (and calendar-handoff splits, which
+        // change the meeting ID mid-recording): pause cotyping and run the
+        // live transcriber against the active meeting folder.
+        recordingStatusObserver = recording.$status
+            .map { status -> UUID? in
+                if case .recording(let meetingID) = status { return meetingID }
+                return nil
+            }
+            .removeDuplicates()
+            .dropFirst()
+            .sink { [weak self] meetingID in
+                self?.meetingRecordingStateDidChange(active: meetingID != nil)
+            }
+        audioMonitorChangeForwarder = audioMonitor.objectWillChange.sink { [weak self] in
+            self?.objectWillChange.send()
+        }
+        calendarObserver = calendar.objectWillChange.sink { [weak self] in
+            self?.objectWillChange.send()
+        }
+        pipeline.onArtifactsWritten = { [weak self] meeting in
+            guard let self else { return }
+            self.outcomeIndex.refresh(meeting: meeting)
+            self.reindexSearchInBackground(meeting)
+            if self.settings.semanticSearchEnabled {
+                self.reindexEmbeddingInBackground(meeting)
+            }
+            self.memoryRoutines.tick()
+            // A parked/failed recovery job may complete after dreaming already
+            // wrote that day's durable marker. Invalidate the generated brief
+            // so the next quiet tick incorporates the recovered artifacts.
+            let calendar = Calendar.current
+            let dayKey = DreamDay.key(for: meeting.startedAt, calendar: calendar)
+            let reportURL = self.dreamStore.reportJSONURL(forDayKey: dayKey)
+            let activationKey = self.settings.dreamingFirstEligibleDayKey ?? dayKey
+            if self.settings.dreamingEnabled,
+               dayKey >= activationKey,
+               FileManager.default.fileExists(atPath: reportURL.path) {
+                do {
+                    try self.dreamStore.invalidateReport(forDayKey: dayKey)
+                } catch {
+                    self.lastError = "Could not refresh the dream for \(dayKey): "
+                        + error.localizedDescription
+                }
+                self.latestDreamReport = self.dreamStore.latestReport()
+                self.dreaming.reconsiderReports()
+            }
+        }
+        if needsSynchronousLibrary {
+            searchIndex.reindexAll(meetings, storage: storage)
+        }
+        if let command = headlessCommand {
+            HeadlessCommandRunner(app: self).run(command)
+            return
+        }
+        // UI tests render against pre-seeded fixtures, not a real audio/Sparkle
+        // session — bail out before any subsystem reaches for the mic, the
+        // process list, or the network.
+        if Self.isUITesting { return }
+        interactive = true
+        RecordingNotifier.shared.bootstrap()
+        applyTrackingSetting()
+        // Crash recovery: re-enqueue any meeting whose processing never
+        // finished — a quit or crash mid-transcription used to lose the job.
+        detector.onMeetingStarted = { [weak self] context in
+            guard let self else { return }
+            switch self.settings.autoRecordMode {
+            case .automatic: self.startRecording(context: context, source: "detector")
+            case .ask: self.notifyMeetingDetected(context)
+            case .manual: RecordingNotifier.shared.invalidateMeetingDetections()
+            }
+        }
+        detector.onMeetingSwitched = { [weak self] context in
+            guard let self else { return }
+            switch self.settings.autoRecordMode {
+            case .automatic:
+                RecordingNotifier.shared.invalidateMeetingDetections()
+                self.recording.splitForCalendarHandoff(context)
+            case .ask:
+                if self.recording.isRecording || self.recording.isStarting {
+                    RecordingNotifier.shared.invalidateMeetingDetections()
+                } else {
+                    self.notifyMeetingDetected(context)
+                }
+            case .manual:
+                RecordingNotifier.shared.invalidateMeetingDetections()
+            }
+        }
+        detector.onMeetingEnded = { [weak self] in
+            RecordingNotifier.shared.invalidateMeetingDetections()
+            self?.stopRecording()
+        }
+        detector.stopDebounce = settings.stopDebounceSeconds
+        detector.calendar = calendar
+        detector.calendarEnabled = settings.calendarDetectionEnabled
+        detector.requireCalendarForBrowser = settings.requireCalendarForBrowser
+        // The mic-in-use signal misses meetings with a muted mic. The audio
+        // monitor is the complementary "a meeting app just started producing
+        // audio output" signal — in automatic mode it auto-records the
+        // recognised meeting bundles; otherwise it surfaces a banner.
+        audioMonitor.start()
+        audioMonitorObserver = audioMonitor.$detectedProcess
+            .compactMap { $0 }
+            .sink { [weak self] process in self?.audioMonitorDetected(process) }
+        detector.start()
+        // Start Sparkle (silent background check). No-op on dev builds and
+        // until the appcast feed URL + public key are configured (RELEASING.md).
+        AppUpdateManager.shared.start()
+        // Resume the wake watcher when the Privacy marker was left enabled.
+        agentAccess.start()
+        screenMemoryAccess.start()
+        applyQuickRecallSetting()
+        applyDailyMemoryExportSetting()
+        applyDayDigestSetting()
+        applyMemoryRoutineSetting()
+        applyDreamingSetting()
+        // First-run check. A genuinely-new user with missing permissions gets
+        // onboarding (windowed — see AppDelegate); the flag persists only when
+        // they explicitly finish or defer the wizard (OnboardingView footer),
+        // so quitting mid-wizard re-opens it next launch. Fully-permissioned
+        // installs that predate the flag are recognised as established right
+        // away and start menu-bar-only next launch.
+        if !UserDefaults.standard.bool(forKey: Self.onboardingShownKey) {
+            if PermissionManager.shared.allGranted {
+                UserDefaults.standard.set(true, forKey: Self.onboardingShownKey)
+            } else {
+                WindowAccess.shared.open("onboarding")
+            }
+        }
+        // Bring optional system-wide automation up if the user left it enabled
+        // and the grants are in place; otherwise it parks itself.
+        dictation.applySettings()
+        cotyping.applySettings()
+        if settings.cotypingEnabled { scheduleCotypingPrewarm() }
+        loadLibraryInBackground()
+    }
+
+    /// Large meeting libraries can take seconds to enumerate, repair, duration-
+    /// probe, and FTS-index. Do that work on a utility executor and publish one
+    /// sorted snapshot when ready instead of blocking the first app window.
+    private func loadLibraryInBackground() {
+        libraryLoadTask?.cancel()
+        let rootURL = storage.rootURL
+        libraryLoadTask = Task { @MainActor [weak self] in
+            let worker = Task.detached(priority: .utility) {
+                guard !Task.isCancelled else { return [Meeting]() }
+                let workerStorage = StorageManager(rootURL: rootURL)
+                return workerStorage.loadMeetings()
+            }
+            let loaded = await withTaskCancellationHandler {
+                await worker.value
+            } onCancel: {
+                worker.cancel()
+            }
+            guard let self, !Task.isCancelled else { return }
+
+            var byID = Dictionary(uniqueKeysWithValues: loaded.map { ($0.id, $0) })
+            // Preserve meetings created while the background scan was running.
+            for meeting in self.meetings { byID[meeting.id] = meeting }
+            let merged = byID.values.sorted { $0.startedAt > $1.startedAt }
+            self.meetings = merged
+            self.libraryReady = true
+            self.outcomeIndex.refresh(meetings: merged)
+            self.pipeline.resumePending(meetings: merged)
+            // Dreaming was deliberately gated while launch recovery rebuilt
+            // the library and durable processing queue. Re-check immediately;
+            // pending pipeline work will keep the downtime gate closed until a
+            // later timer tick.
+            self.dreaming.tick()
+            self.reindexLibraryInBackground(merged)
+            self.libraryLoadTask = nil
+            if let pending = self.pendingRecordingStart {
+                self.pendingRecordingStart = nil
+                self.lastError = nil
+                self.startRecording(context: pending.context, source: pending.source)
+            }
+        }
+    }
+
+    private func reindexLibraryInBackground(_ meetings: [Meeting]) {
+        let worker = searchIndexWorkQueue
+        Task { await worker.enqueue(meetings) }
+    }
+
+    private func reindexSearchInBackground(_ meeting: Meeting) {
+        let worker = searchIndexWorkQueue
+        Task { await worker.enqueue(meeting) }
+    }
+
+    private func reindexEmbeddingInBackground(_ meeting: Meeting) {
+        embeddingIndexTasks.removeValue(forKey: meeting.id)?.task.cancel()
+        let token = UUID()
+        let task = Task { @MainActor [weak self] in
+            guard let self, !Task.isCancelled else { return }
+            try? await self.embeddingIndex.index(meeting)
+            guard self.embeddingIndexTasks[meeting.id]?.token == token else { return }
+            self.embeddingIndexTasks.removeValue(forKey: meeting.id)
+        }
+        embeddingIndexTasks[meeting.id] = (token, task)
+    }
+
+    /// Index rows are reconstructible, but a failed SQLite cleanup must not
+    /// become permanent once the meeting folder disappears. Retry both stores
+    /// with bounded backoff for the rest of the session; durable tombstones and
+    /// each index's startup reconciliation cover interruption or app exit.
+    private func scheduleIndexCleanup(_ meetingID: Meeting.ID) {
+        indexCleanupTasks.removeValue(forKey: meetingID)?.task.cancel()
+        let token = UUID()
+        let worker = searchIndexWorkQueue
+        let task = Task { @MainActor [weak self] in
+            guard let self else { return }
+            var embeddingClean = false
+            var searchClean = false
+            var attempt = 0
+
+            while !Task.isCancelled {
+                let result = await worker.remove(meetingID)
+                embeddingClean = embeddingClean || result.embedding
+                searchClean = searchClean || result.search
+                guard !Task.isCancelled,
+                      self.indexCleanupTasks[meetingID]?.token == token else { return }
+                if embeddingClean && searchClean {
+                    self.indexCleanupTasks.removeValue(forKey: meetingID)
+                    return
+                }
+
+                attempt += 1
+                if attempt.isMultiple(of: 4) {
+                    let reconciled = await worker.reconcileDeletedMeetings()
+                    embeddingClean = embeddingClean || reconciled.embedding
+                    searchClean = searchClean || reconciled.search
+                }
+                let delayMilliseconds = min(30_000, 250 * (1 << min(attempt, 7)))
+                try? await Task.sleep(for: .milliseconds(delayMilliseconds))
+            }
+        }
+        indexCleanupTasks[meetingID] = (token, task)
+    }
+
+    private enum CotypingRuntimeOperation {
+        case unload
+        case prewarm
+    }
+
+    private func scheduleCotypingRuntimeUnload() {
+        scheduleCotypingRuntimeOperation(.unload)
+    }
+
+    private func scheduleCotypingPrewarm() {
+        scheduleCotypingRuntimeOperation(.prewarm)
+    }
+
+    /// Queue lifecycle changes instead of launching independent tasks. An
+    /// unload intentionally runs even if a later operation cancels its task:
+    /// the later operation awaits it before loading another model.
+    private func scheduleCotypingRuntimeOperation(_ operation: CotypingRuntimeOperation) {
+        let previous = cotypingRuntimeTask
+        previous?.cancel()
+        let taskID = UUID()
+        cotypingRuntimeTaskID = taskID
+        cotypingRuntimeTask = Task { @MainActor [weak self] in
+            if let previous { await previous.value }
+            guard let self else { return }
+            defer {
+                if self.cotypingRuntimeTaskID == taskID {
+                    self.cotypingRuntimeTask = nil
+                    self.cotypingRuntimeTaskID = nil
+                }
+            }
+            switch operation {
+            case .unload:
+                await self.cotypingEngine.unload()
+            case .prewarm:
+                guard !Task.isCancelled, self.settings.cotypingEnabled else { return }
+                await self.cotypingEngine.prewarm()
+            }
+        }
+    }
+
+    func prepareForTermination() async {
+        if let terminationCleanupTask {
+            await terminationCleanupTask.value
+            return
+        }
+        let task = Task { @MainActor in
+            interactive = false
+            settingsStore.flush()
+            let pendingCotypingRuntimeTask = self.cotypingRuntimeTask
+            pendingCotypingRuntimeTask?.cancel()
+            await pendingCotypingRuntimeTask?.value
+            self.cotypingRuntimeTask = nil
+            cotypingRuntimeTaskID = nil
+            libraryLoadTask?.cancel()
+            libraryLoadTask = nil
+            for entry in embeddingIndexTasks.values { entry.task.cancel() }
+            embeddingIndexTasks.removeAll()
+            for entry in indexCleanupTasks.values { entry.task.cancel() }
+            indexCleanupTasks.removeAll()
+            await searchIndexWorkQueue.stop()
+            pendingRecordingStart = nil
+            recording.prepareForTermination()
+            detector.stop()
+            audioMonitor.stop()
+            sampler.stop()
+            screenshots.stop()
+            quickRecallHotKey.stop()
+            dailyMemoryExportScheduler.stop()
+            memoryRoutines.stop()
+            dreaming.stop()
+            chat.stop()
+            dictation.stop()
+            cotyping.stop()
+            await cotypingLearning.flushPersistence()
+            await CotypingStatsStore.shared.flushPersistence()
+            await agentSessions.shutdownAll()
+            await cotypingEngine.unload()
+            await LlamaServer.shared.stop()
+            await LlamaServer.embedder.stop()
+            await LlamaServer.cotyping.stop()
+            await GraniteSpeechEngine.shared.shutdown()
+        }
+        terminationCleanupTask = task
+        await task.value
+    }
+
+    // MARK: - Detection → recording glue
+
+    /// Builds a detection context for a user-initiated recording on `detectedApp`
+    /// (nil → manual), folding in the active calendar event when calendar
+    /// detection is enabled and authorized — so the menu / command / banner
+    /// entry points get calendar titling too.
+    func recordingContext(for detectedApp: MeetingDetector.DetectedApp?) -> MeetingDetectionContext? {
+        guard let detectedApp else { return nil }
+        let event = (settings.calendarDetectionEnabled && calendar.hasAccess)
+            ? calendar.activeCandidate(now: Date()) : nil
+        return MeetingDetectionContext(
+            detectedApp: detectedApp,
+            calendarEvent: event,
+            confidence: MeetingMatcher.confidence(hasApp: true, hasCalendar: event != nil),
+            reason: "user")
+    }
+
+    /// A user can explicitly start the scheduled event from Today's upcoming
+    /// meeting card before automatic detection fires. Attach the calendar
+    /// metadata immediately; capture system audio too when a matching meeting
+    /// app is already visible, otherwise RecordingController safely records
+    /// the microphone when no meeting app is visible yet.
+    func recordingContext(for calendarEvent: CalendarMeetingCandidate) -> MeetingDetectionContext {
+        let detectedApp = detector.activeApp ?? MeetingDetector.visibleBrowserMeeting()
+        return MeetingDetectionContext(
+            detectedApp: detectedApp,
+            calendarEvent: calendarEvent,
+            confidence: MeetingMatcher.confidence(
+                hasApp: detectedApp != nil,
+                hasCalendar: true),
+            reason: "today-upcoming")
+    }
+
+    /// `AudioSourceMonitor` saw an app newly start producing output. Auto-record
+    /// in automatic mode only for high-confidence native meeting output or a
+    /// calendar-backed native app; leave broader chat apps as banner/detector
+    /// candidates so notification sounds cannot start recordings.
+    private func audioMonitorDetected(_ process: AudioProcess) {
+        guard !recording.isRecording, !recording.isStarting else { return }
+        guard settings.autoRecordMode == .automatic, let bundleID = process.bundleID else { return }
+        let calendarEvent = (settings.calendarDetectionEnabled && calendar.hasAccess)
+            ? calendar.activeCandidate(now: Date()) : nil
+        if let name = MeetingDetector.knownApps[bundleID] {
+            guard MeetingDetector.shouldAutoRecordNativeAudioMonitor(
+                bundleID: bundleID,
+                calendarBacked: calendarEvent != nil) else {
+                return
+            }
+            let detected = MeetingDetector.DetectedApp(name: name, bundleID: bundleID, pid: process.id)
+            startRecording(context: detectionContext(detected, calendarEvent), source: "audio-monitor")
+            return
+        }
+        guard let hostBundleID = MeetingDetector.hostBrowserBundleID(forAudioBundleID: bundleID) else { return }
+        // The browser is already producing output (the monitor fired on it), so a
+        // window-title match OR an active calendar meeting link is enough — the
+        // latter catches a generic-title Google Meet the title check misses.
+        let titleMatches = MeetingDetector.visibleBrowserMeeting()?.bundleID == hostBundleID
+        let calendarBacked = calendarEvent?.meetingURL != nil
+        guard MeetingMatcher.browserCountsAsMeeting(
+            titleMatchesMarker: titleMatches, hasOutputAudio: true,
+            calendarBacked: calendarBacked, requireCalendarForBrowser: settings.requireCalendarForBrowser)
+        else { return }
+        let name = NSRunningApplication.runningApplications(withBundleIdentifier: hostBundleID)
+            .first?.localizedName ?? "Browser"
+        let detected = MeetingDetector.DetectedApp(name: name, bundleID: hostBundleID, pid: process.id)
+        startRecording(context: detectionContext(detected, calendarEvent), source: "audio-monitor")
+    }
+
+    private func detectionContext(_ app: MeetingDetector.DetectedApp,
+                                  _ event: CalendarMeetingCandidate?) -> MeetingDetectionContext {
+        MeetingDetectionContext(
+            detectedApp: app, calendarEvent: event,
+            confidence: MeetingMatcher.confidence(hasApp: true, hasCalendar: event != nil),
+            reason: "audio-monitor")
+    }
+
+    private func notifyMeetingDetected(_ context: MeetingDetectionContext) {
+        lastError = nil
+        let title = MeetingMatcher.recordingTitle(
+            calendarTitle: context.calendarEvent?.title,
+            useCalendarTitles: settings.useCalendarTitles,
+            appName: context.detectedApp?.name)
+        RecordingNotifier.shared.meetingDetected(title: title) { [weak self] in
+            guard let self,
+                  self.settings.autoRecordMode == .ask,
+                  !self.recording.isRecording,
+                  !self.recording.isStarting else { return }
+            self.startRecording(context: context, source: "notification")
+        }
+    }
+
+    // MARK: - Library operations
+
+    func reprocess(_ meeting: Meeting, transcribe: Bool, summarize: Bool) {
+        pipeline.enqueue(meeting, transcribe: transcribe, summarize: summarize)
+    }
+
+    /// Row-level recovery intentionally retries the complete durable pipeline.
+    /// It is the safest default when a compact row cannot know which earlier
+    /// artifact came from the failed attempt; the detail Process menu still
+    /// offers narrower transcription- or summary-only choices.
+    func retryProcessing(_ meeting: Meeting) {
+        reprocess(meeting, transcribe: true, summarize: true)
+    }
+
+    func saveTranscript(_ transcript: Transcript, for meeting: Meeting) throws {
+        try pipeline.saveTranscript(transcript, for: meeting)
+        reindexSearchInBackground(meeting)
+        if settings.semanticSearchEnabled {
+            reindexEmbeddingInBackground(meeting)
+        }
+        objectWillChange.send()
+    }
+
+    func speakerNameHints(for meeting: Meeting) -> [String] {
+        let fallbackDuration = max(meeting.recordedDuration ?? 60, 60)
+        let fallbackEnd = meeting.startedAt.addingTimeInterval(fallbackDuration)
+        var end = meeting.endedAt.map { max($0, meeting.startedAt) } ?? fallbackEnd
+        if end <= meeting.startedAt { end = fallbackEnd }
+        let ocr = activityStore.ocrText(from: meeting.startedAt, to: end, maxChars: 12_000)
+        return SpeakerNameHintExtractor.hints(
+            calendarNames: meeting.participantNameHints ?? [],
+            ocrText: ocr)
+    }
+
+    func applyTrackingSetting() {
+        sampler.excludedApps = { [weak self] in self?.settings.excludedAppList ?? [] }
+        if settings.trackingEnabled { sampler.start() } else { sampler.stop() }
+        screenshots.restart()
+    }
+
+    func applyQuickRecallSetting() {
+        let registered = quickRecallHotKey.setEnabled(settings.quickRecallEnabled)
+        if settings.quickRecallEnabled, !registered {
+            settings.quickRecallEnabled = false
+            lastError = "The Ask shortcut could not register \(QuickRecallHotKeyController.shortcutLabel). Another app may already use it."
+        }
+    }
+
+    func applyDailyMemoryExportSetting() {
+        let snapshot = settings
+        let destinationID = snapshot.dailyMemoryExportFolder.isEmpty
+            ? ""
+            : "\(snapshot.dailyMemoryExportFolder)|\(snapshot.dailyMemoryExportFormat.rawValue)"
+        let configuration = DailyMemoryExportScheduler.Configuration(
+            enabled: snapshot.dailyMemoryExportEnabled,
+            hour: snapshot.dailyMemoryExportHour,
+            destinationID: destinationID)
+        let storageRoot = storage.rootURL
+        let destinationPath = snapshot.dailyMemoryExportFolder
+        let exportKind: DailyMemoryExportKind = switch snapshot.dailyMemoryExportFormat {
+        case .markdown: .markdown
+        case .obsidian: .obsidian
+        case .logseq: .logseq
+        }
+        dailyMemoryExportScheduler.configure(configuration) { day in
+            let destination = URL(
+                fileURLWithPath: destinationPath,
+                isDirectory: true)
+            try Task.checkCancellation()
+            let service = DailyMemoryExportService(
+                source: FileDailyMemoryExportSource(root: storageRoot),
+                calendar: .current)
+            _ = try service.export(
+                day: day,
+                configuration: DailyMemoryExportConfiguration(
+                    destinationDirectory: destination,
+                    format: exportKind))
+        } onError: { [weak self] message in
+            self?.lastError = message
+        }
+    }
+
+    func applyDayDigestSetting() {
+        let snapshot = settings
+        let storageRoot = storage.rootURL
+        dayDigestScheduler.configure(
+            DayDigestScheduler.Configuration(
+                enabled: snapshot.dayDigestAutoEnabled,
+                hour: snapshot.dayDigestHour),
+            digestModifiedAt: { day in
+                let name = DreamDay.key(for: day)
+                let url = storageRoot.appendingPathComponent("journal/\(name).md")
+                let attributes = try? FileManager.default.attributesOfItem(atPath: url.path)
+                return attributes?[.modificationDate] as? Date
+            },
+            latestEvidenceAt: { [weak self] day in
+                guard let self else { return nil }
+                let meetingEnd = self.meetings
+                    .filter { Calendar.current.isDate($0.startedAt, inSameDayAs: day) }
+                    .compactMap(\.endedAt)
+                    .max()
+                return [self.activityStore.latestEvidenceAt(on: day), meetingEnd]
+                    .compactMap { $0 }
+                    .max()
+            },
+            canRun: { [weak self] in
+                guard let self, self.libraryReady else { return false }
+                let cotypingGenerating: Bool
+                if case .generating = self.cotyping.state {
+                    cotypingGenerating = true
+                } else {
+                    cotypingGenerating = false
+                }
+                return !self.recording.isRecording
+                    && !self.recording.isStarting
+                    && !self.dictation.isStarting
+                    && !self.dictation.state.isWorking
+                    && !cotypingGenerating
+                    && !self.pipeline.hasActiveWork
+            },
+            generate: { [weak self] day in
+                guard let self else {
+                    throw TextEngineError.unavailable("LokalBot is shutting down.")
+                }
+                let blocks = self.activityStore.blocks(on: day)
+                let finished = self.meetings.filter {
+                    Calendar.current.isDate($0.startedAt, inSameDayAs: day)
+                        && $0.endedAt != nil
+                }
+                let screenContexts = self.activityStore.screenContexts(on: day)
+                guard !blocks.isEmpty || !finished.isEmpty || !screenContexts.isEmpty else {
+                    return false
+                }
+                let result = try await self.pipeline.generateDayDigest(
+                    for: day, blocks: blocks, meetings: finished,
+                    screenContexts: screenContexts,
+                    config: self.settings)
+                if let warning = result.summaryWarning {
+                    self.lastError = warning
+                }
+                return true
+            },
+            onError: { [weak self] message in
+                self?.lastError = message
+            })
+    }
+
+    func applyMemoryRoutineSetting() {
+        let snapshot = settings
+        memoryRoutines.configure(MemoryRoutineScheduler.Configuration(
+            enabled: snapshot.memoryRoutinesEnabled,
+            destinationPath: snapshot.memoryRoutineFolder,
+            kinds: snapshot.enabledMemoryRoutines,
+            hour: snapshot.memoryRoutineHour,
+            weekday: snapshot.memoryRoutineWeekday
+        )) { [weak self] message in
+            self?.lastError = message
+        }
+    }
+
+    func applyDreamingSetting() {
+        let snapshot = settings
+
+        // Persist the beginning of each opt-in period. This keeps catch-up
+        // bounded instead of interpreting a first launch as permission to
+        // backfill every historical workday.
+        if snapshot.dreamingEnabled,
+           snapshot.dreamingFirstEligibleDayKey?.isEmpty != false {
+            let calendar = Calendar.current
+            var updated = snapshot
+            updated.dreamingFirstEligibleDayKey = DreamDay.key(
+                for: DreamScheduler.previousDay(of: Date(), calendar: calendar),
+                calendar: calendar)
+            settings = updated
+            return
+        }
+
+        if !snapshot.dreamingEnabled,
+           snapshot.dreamingFirstEligibleDayKey != nil {
+            var updated = snapshot
+            updated.dreamingFirstEligibleDayKey = nil
+            settings = updated
+            return
+        }
+
+        let storageRoot = storage.rootURL
+        dreaming.configure(
+            DreamScheduler.Configuration(
+                enabled: snapshot.dreamingEnabled,
+                hour: snapshot.dreamingHour,
+                firstEligibleDayKey: snapshot.dreamingFirstEligibleDayKey ?? ""),
+            hasReport: { dayKey in
+                DreamStore(root: storageRoot).hasReport(forDayKey: dayKey)
+            },
+            canRun: { [weak self] in
+                guard let self else { return false }
+                guard self.libraryReady else { return false }
+                let cotypingGenerating: Bool
+                if case .generating = self.cotyping.state {
+                    cotypingGenerating = true
+                } else {
+                    cotypingGenerating = false
+                }
+                let lokalBotIsIdle = !self.recording.isRecording
+                    && !self.recording.isStarting
+                    && !self.dictation.isStarting
+                    && !self.dictation.state.isWorking
+                    && !cotypingGenerating
+                    && !self.pipeline.hasActiveWork
+                guard lokalBotIsIdle else { return false }
+                guard DreamScheduler.powerAllowsDreaming(
+                    isOnBattery: PowerSourceMonitor.currentlyOnBattery(),
+                    isLowPower: ProcessInfo.processInfo.isLowPowerModeEnabled) else {
+                    return false
+                }
+                let userIdleSeconds = CGEventSource.secondsSinceLastEventType(
+                    .combinedSessionState, eventType: CGEventType(rawValue: ~0)!)
+                return DreamScheduler.isSystemIdle(for: userIdleSeconds)
+            },
+            dream: { [weak self] target in
+                guard let self else {
+                    throw TextEngineError.unavailable("LokalBot is shutting down.")
+                }
+                let service = DreamService(
+                    storageRoot: storageRoot,
+                    makeEngine: { [weak self] in
+                        guard let self else {
+                            throw TextEngineError.unavailable("LokalBot is shutting down.")
+                        }
+                        let engineSettings = self.settings
+                        let engine = try await self.pipeline.makeTextEngine(
+                            engineSettings,
+                            priority: .background,
+                            purpose: "dreaming")
+                        return (engine, DreamInferenceProvenance(settings: engineSettings))
+                    })
+                let report = try await service.dream(target: target)
+                self.latestDreamReport = report
+                self.refreshDreamMemory()
+            },
+            onError: { [weak self] message in
+                self?.lastError = message
+            })
+    }
+
+    func setDreamingEnabled(_ enabled: Bool) {
+        guard settings.dreamingEnabled != enabled else { return }
+
+        var updated = settings
+        updated.dreamingEnabled = enabled
+        if enabled {
+            let calendar = Calendar.current
+            updated.dreamingFirstEligibleDayKey = DreamDay.key(
+                for: DreamScheduler.previousDay(of: Date(), calendar: calendar),
+                calendar: calendar)
+        } else {
+            updated.dreamingFirstEligibleDayKey = nil
+        }
+        settings = updated
+    }
+
+    func dreamNow() {
+        guard libraryReady else {
+            lastError = "Preparing your meeting library; dreaming will be available when it is ready."
+            return
+        }
+        dreaming.dreamNow()
+    }
+
+    func refreshDreamMemory() {
+        do {
+            dreamMemory = try dreamStore.loadMemory()
+        } catch {
+            dreamMemory = nil
+            lastError = "Could not load dream memory: \(error.localizedDescription)"
+        }
+    }
+
+    func setDreamMemoryPinned(_ pinned: Bool, for entry: DreamMemoryEntry) {
+        guard !dreaming.isDreaming else {
+            lastError = "Wait for the current dream to finish before changing pins."
+            return
+        }
+        do {
+            guard let updated = try dreamStore.setPinned(pinned, for: entry) else {
+                refreshDreamMemory()
+                lastError = "That dream memory item is no longer available."
+                return
+            }
+            dreamMemory = updated
+        } catch {
+            lastError = "Could not update dream memory: \(error.localizedDescription)"
+        }
+    }
+
+    func restartMemoryCapture() {
+        sampler.stop()
+        screenshots.stop()
+        applyTrackingSetting()
+        PermissionManager.shared.refresh()
+    }
+
+    /// Search hit → open the meeting; transcript hits seek the player.
+    func openSearchHit(_ hit: SearchIndex.Hit) {
+        if hit.kind == .segment {
+            pendingSeek = hit.start
+        }
+        openMeeting(hit.meetingID)
+    }
+
+    /// Screen search/citation hit → open Timeline at the exact captured frame.
+    func openScreenSnapshot(_ snapshotID: Int64) {
+        pendingScreenSnapshotID = snapshotID
+        selectedMeetingIDs = []
+        navSection = .timeline
+    }
+
+    /// Chat citation marker → open the cited meeting; timed markers seek the player.
+    func openCitation(_ citation: ChatCitation) {
+        guard let meeting = ((try? SessionLookup.find(id: citation.meetingID, in: meetings)) ?? nil) else { return }
+        if let seconds = citation.seconds {
+            pendingSeek = seconds
+        }
+        openMeeting(meeting.id)
+    }
+
+    /// Permanently removes meetings: audio folder, list entry, both indexes.
+    func deleteMeetings(_ ids: Set<Meeting.ID>) {
+        var deletedIDs: Set<Meeting.ID> = []
+        for meeting in meetings where ids.contains(meeting.id) {
+            do {
+                try storage.deleteMeeting(meeting)
+                embeddingIndexTasks.removeValue(forKey: meeting.id)?.task.cancel()
+                deletedMeetingIDs.insert(meeting.id)
+                cachedSearchIndex?.noteDeletion(meeting.id)
+                cachedEmbeddingIndex?.noteDeletion(meeting.id)
+                scheduleIndexCleanup(meeting.id)
+                deletedIDs.insert(meeting.id)
+            } catch {
+                lastError = "Could not delete \(meeting.title): \(error.localizedDescription)"
+            }
+        }
+        meetings.removeAll { deletedIDs.contains($0.id) }
+        selectedMeetingIDs.subtract(deletedIDs)
+        outcomeIndex.refresh(meetings: meetings)
+    }
+}
